display full weather info

+++ src/weather.h
@@ -14,7 +14,7 @@ struct Weather {
     float rainFall;
     uint8_t precip;
     char windDirection[3];
-    uint8_t windSpeed;
+    float windSpeed;
     uint8_t windRank;
     uint8_t humidity;
 };
